Improve docs on using `"externalHelpers": true` (#300)
When you set `"externalHelpers": true`, SWC will inject ESM import
statements to the top of the file, regardless of if this module is
commonjs or ESM.

This can mess up commonjs files because they'd now have mixed import
syntaxes.

This can be fixed by setting `"isModule": "unkown"`.

So i've added a warning explaining that and pointing the user to the
right direction.

Resloves: #299